Changed files, tested in scene
Added some gameobjects to test the flamethrower, made player point to mouse

+++ Assets/Scripts/ComboValues.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 public class ComboValues : MonoBehaviour
 {
     public List<float> mults = new List<float>();
-    public List<float> threshholds = new List<float>();
+    public List<int> threshholds = new List<int>();
 }
